[Feat] - 공통 가게 - 즐겨찾기 API 5개 추가
