+++ app/src/main/java/com/zion830/threedollars/datasource/FavoriteMyFolderDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.zion830.threedollars.datasource
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse
+import com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse
 import com.zion830.threedollars.di.NetworkModule
 import com.zion830.threedollars.network.ServiceApi
 import com.zion830.threedollars.utils.getErrorMessage
