+++ app/src/main/java/com/zion830/threedollars/ui/favorite/FavoriteMyFolderRecyclerAdapter.kt
@@ -5,7 +5,7 @@ import androidx.core.view.isVisible
 import androidx.paging.PagingDataAdapter
 import com.zion830.threedollars.R
 import com.zion830.threedollars.databinding.ItemFavoritePageBinding
-import com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse
+import com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse
 import zion830.com.common.base.BaseDiffUtilCallback
 import zion830.com.common.base.BaseViewHolder
 import zion830.com.common.listener.OnItemClickListener
