+++ app/src/main/java/com/zion830/threedollars/ui/favorite/viewer/FavoriteViewerActivity.kt
@@ -9,7 +9,7 @@ import com.zion830.threedollars.Constants
 import com.zion830.threedollars.GlobalApplication
 import com.zion830.threedollars.R
 import com.zion830.threedollars.databinding.ActivityFavoriteViewerBinding
-import com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse
+import com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse
 import com.zion830.threedollars.ui.food_truck_store_detail.FoodTruckStoreDetailActivity
 import com.zion830.threedollars.ui.login.dialog.LoginRequestDialog
 import com.zion830.threedollars.ui.login.name.InputNameActivity
