+++ app/src/main/res/layout/activity_favorite_viewer.xml
@@ -13,7 +13,7 @@
 
         <variable
             name="item"
-            type="com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse" />
+            type="com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
