+++ app/src/main/res/layout/item_favorite_page.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse.MyFavoriteFolderFavoriteModel" />
+            type="com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
