+++ app/src/main/res/layout/item_mypage_favorite.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="item"
-            type="com.zion830.threedollars.datasource.model.v2.response.favorite.MyFavoriteFolderResponse.MyFavoriteFolderFavoriteModel" />
+            type="com.zion830.threedollars.datasource.model.v4.favorite.MyFavoriteFolderResponse" />
     </data>
 
     <LinearLayout
